Delete HintPath from HypnoLogLib.csproj
This should solve problems when using HypnoLogLib.csproj as project in
some other solution.
The issue is with HintPath nodes pointing to specific nuget package
location relative to the project location. This path is not correct in
case HypnoLogLib is located in another folder (as happens when using it
as git submodule in other project).