Redirect console output in intialization
RedirectConsoleOutput is an argument in initialization and not method
becouse when redirecting console output the HL server
must be on and workong (= initializtion was succesfull),
we do not want to redirect console output if initialization failed.
This is also because initialization is non-blocking
(and we want it to be so).

Conflicts:
	HypnoLogLib/HypnoLog.cs