Add Reference HintPath to csproj
This should solve problems finding nuget packages dll files when
HypnoLogLib used as submodule.

The problem was:
When HypnoLogLib used as its own solution, packages folder located on
same level as the .sln file so hint path is ../packages/
When HypnoLogLib used as submodule and included in other solution,
packages folder located on the level of main repo .sln file so
hint path is ../../packages/

Solution is to add single hint path (2 hint paths is not supported) to
point on the packages folder which now included in repo.