Allow setting type when logging
Add `type` argument to all logging methods, so user can explicitly set the
logging data type (and then control the visualization).

For this, refactor how we determine the implicit type, when not given by user.

Also, refactor examples, to contain "basic" and "advanced".
Make the basic work with the minimal-required examples,
move all the experimental stuff to advanced (WIP: need to fix tags, and other
features).